[#66] 테스팅 파일 병합
